+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 agp = "8.7.3"
-datastorePreferences = "1.1.3"
+datastorePreferences = "1.1.4"
 hiltAndroidCompiler = "2.55"
 kotlin = "2.0.0"
 coreKtx = "1.15.0"
@@ -10,13 +10,17 @@ espressoCore = "3.6.1"
 kotlinxSerializationJson = "1.8.0"
 lifecycleRuntimeKtx = "2.8.7"
 activityCompose = "1.10.1"
-composeBom = "2025.03.00"
-lifecycleViewmodelCompose = "2.9.0-alpha12"
-material = "1.13.0-alpha11"
+composeBom = "2025.03.01"
+lifecycleViewmodelCompose = "2.9.0-alpha13"
+material = "1.13.0-alpha12"
 materialIconsExtended = "1.7.8"
 onnxruntimeAndroid = "1.21.0"
 onnxruntimeExtensionsAndroid = "0.13.0"
-runtimeLivedata = "1.8.0-rc01"
+runtimeLivedata = "1.8.0-rc02"
+material3WindowSizeClassAndroid = "1.3.1"
+adaptiveAndroid = "1.1.0"
+window = "1.4.0-rc01"
+material3AdaptiveNavigationSuiteAndroid = "1.3.1"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -24,6 +28,7 @@ androidx-datastore-preferences = { module = "androidx.datastore:datastore-prefer
 androidx-lifecycle-viewmodel-compose = { module = "androidx.lifecycle:lifecycle-viewmodel-compose", version.ref = "lifecycleViewmodelCompose" }
 androidx-material-icons-extended = { module = "androidx.compose.material:material-icons-extended", version.ref = "materialIconsExtended" }
 androidx-runtime-livedata = { module = "androidx.compose.runtime:runtime-livedata", version.ref = "runtimeLivedata" }
+androidx-window = { module = "androidx.window:window", version.ref = "window" }
 hilt-android = { module = "com.google.dagger:hilt-android", version.ref = "hiltAndroidCompiler" }
 hilt-android-compiler = { module = "com.google.dagger:hilt-android-compiler", version.ref = "hiltAndroidCompiler" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
@@ -43,6 +48,9 @@ kotlinx-serialization-json = { module = "org.jetbrains.kotlinx:kotlinx-serializa
 material = { module = "com.google.android.material:material", version.ref = "material" }
 onnxruntime-android = { module = "com.microsoft.onnxruntime:onnxruntime-android", version.ref = "onnxruntimeAndroid" }
 onnxruntime-extensions-android = { module = "com.microsoft.onnxruntime:onnxruntime-extensions-android", version.ref = "onnxruntimeExtensionsAndroid" }
+androidx-material3-window-size = { group = "androidx.compose.material3", name = "material3-window-size-class-android", version.ref = "material3WindowSizeClassAndroid" }
+androidx-adaptive-android = { group = "androidx.compose.material3.adaptive", name = "adaptive-android", version.ref = "adaptiveAndroid" }
+androidx-material3-adaptive-navigation-suite-android = { group = "androidx.compose.material3", name = "material3-adaptive-navigation-suite-android", version.ref = "material3AdaptiveNavigationSuiteAndroid" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
